Commit initial type/provider/class
Yes, this is ugly. I don't think I want a resource called
'munki_package' - it's better to have a package resource with a munki
provider.  The problem is that I wanted to create a catalog parameter
for the main package type, but didn't want to throw the current package
type into a module and modify it just for that.  So I did this.  It
works, it's basic, and it's definitely not to be used in production.
Consider this 'proof of concept'.  Please.